% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_title.R
\name{addTitle}
\alias{addTitle}
\title{Add a title to a workbook core information, which will display in the file information tab in Excel.}
\usage{
addTitle(wb, title, overwrite = FALSE)
}
\arguments{
\item{wb}{A Workbook object.}

\item{title}{A string containing the title for the Workbook}

\item{overwrite}{boolean specifying if the title should overwrite any existing title. Defaults to FALSE}
}
\description{
Add a title to a workbook core information, which will display in the file information tab in Excel.
}
