% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_row_styles.R
\name{copyRowStyles}
\alias{copyRowStyles}
\title{Copy the formatting of a row to another row for a defined selection of columns}
\usage{
copyRowStyles(wb, sheet, rowFrom, rowTo, cols)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The name of the worksheet to write to.}

\item{rowFrom}{Integer showing the row number in the excel workbook to 
copy the formatting from}

\item{rowTo}{An integer (or vector of integers) defining the row number 
in the excel workbook for the format to be copied to}

\item{cols}{A vector of one or more integer indicating the cols 
in the excel workbook to copy the format from}
}
\description{
Copy the formatting of a row to another row for a defined selection of columns
}
