% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superscript_notes.R
\name{superscriptNotes}
\alias{superscriptNotes}
\title{Convert footnote annotations using the format [note x] to superscript}
\usage{
superscriptNotes(wb)
}
\arguments{
\item{wb}{A Workbook object containing one or more worksheets.}
}
\description{
Convert footnote annotations using the format [note x] to superscript
}
