% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_template.R
\name{writeDataTableTemplate}
\alias{writeDataTableTemplate}
\title{Write to a worksheet template and format as an Excel table}
\usage{
writeDataTableTemplate(
  wb,
  sheet,
  x,
  startRow,
  tableName = NULL,
  startCol = 1,
  colNames = TRUE,
  rowNames = FALSE,
  keepNA = openxlsx_getOp("keepNA", FALSE),
  na.string = openxlsx_getOp("na.string")
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{x}{A dataframe}

\item{startRow}{A vector specifying the starting row to write df}

\item{tableName}{name of table in workbook. The table name must be unique.}

\item{startCol}{A vector specifying the starting column to write df}

\item{colNames}{If TRUE, column names of x are written.}

\item{rowNames}{If TRUE, row names of x are written.}

\item{keepNA}{If TRUE, NA values are converted to #N/A (or na.string, if not NULL) in Excel, else NA cells will be empty.}

\item{na.string}{If not NULL, and if keepNA is TRUE, NA values are converted to this string in Excel.}
}
\description{
Write to a worksheet template and format as an Excel table
}
