% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReplaceData.R
\name{writeReplaceDataTable}
\alias{writeReplaceDataTable}
\title{Write to an Excel table replacing NA values in a custom way}
\usage{
writeReplaceDataTable(
  wb,
  sheet,
  x,
  colNames = TRUE,
  tableName = NULL,
  startCol = 1,
  startRow = 1,
  keepNA = TRUE,
  na.string = "[z]",
  ...
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{x}{A dataframe}

\item{colNames}{If TRUE, column names of x are written.}

\item{tableName}{name of table in workbook. The table name must be unique.}

\item{startCol}{A vector specifying the starting column to write df}

\item{startRow}{A vector specifying the starting row to write df}

\item{keepNA}{If TRUE, NA values are converted to #N/A 
(or na.string, if not NULL) in Excel. If FALSE, NA cells will be empty. 
A list of column numbers can also be passed to this argument; in this case
NA values in the first columns in the list will be converted to the 
first string provided to na.string, the second to the second string, etc.}

\item{na.string}{If keepNA is TRUE, all NA values are converted to this string in Excel.
If keepNA is a list of column numbers a vector of values can be provided to this argument, 
NA values in the specified columns are converted to their corresponding string}
}
\description{
Write to an Excel table replacing NA values in a custom way
}
