% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReplaceMultiple.R
\name{writeReplaceMultiple}
\alias{writeReplaceMultiple}
\title{Write to an Excel table replacing NA values in a custom way}
\usage{
writeReplaceMultiple(
  wb,
  sheet,
  x,
  colNames = TRUE,
  tableName = NULL,
  startCol = 1,
  startRow = 1,
  keepNA = TRUE,
  na.string = "[z]",
  keepZero = TRUE,
  zero.string = NULL,
  withFilter = FALSE,
  ...
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{x}{A dataframe}

\item{colNames}{If TRUE, column names of x are written.}

\item{tableName}{name of table in workbook. The table name must be unique.}

\item{startCol}{A vector specifying the starting column to write df}

\item{startRow}{A vector specifying the starting row to write df}

\item{keepNA}{If TRUE, NA values are converted to #N/A 
(or na.string, if not NULL) in Excel. If FALSE, NA cells will be empty.}

\item{na.string}{If keepNA is TRUE, all NA values are converted to this string in Excel.}

\item{keepZero}{If FALSE, 0 values are converted to the zero.string. If TRUE, they will remain as 0.}

\item{zero.string}{If keepZero is FALSE, all zero values are converted to this string in Excel}

\item{withFilter}{whether to include filters on the table or not. Defaults to FALSE}

\item{...}{Additional arguments to be passed to writeDataTable}
}
\description{
Write to an Excel table replacing NA values in a custom way
}
