% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_column_styles.R
\name{copyColumnStyles}
\alias{copyColumnStyles}
\title{Copy the formatting of a column to another column for a defined consecutive selection of rows}
\usage{
copyColumnStyles(wb, sheet, colFrom, colTo, startRow, endRow)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{colFrom}{Integer showing the column number in the excel workbook of which the format is to be copied}

\item{colTo}{An Integer defining the column number in the excel workbook for the format to be copied to}

\item{startRow}{An integer indicating the row in the excel workbook to start copying the format}

\item{endRow}{An integer indicating the row in the excel workbook to stop copying the format}
}
\description{
Copy the formatting of a column to another column for a defined consecutive selection of rows
}
